Create post layout, restructure
