+++ pages/posts.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Layout from "../components/layout";
+import Layout from "../layouts/main";
 
 const Home = () => {
   return <Layout title="Posts"></Layout>;
